refactor: product 조회 로직 CategoryController에서 ProductController로 변경
