+++ src/main/java/com/flab/daitso/controller/CategoryController.java
@@ -1,14 +1,11 @@
 package com.flab.daitso.controller;
 
 import com.flab.daitso.dto.product.Category;
-import com.flab.daitso.dto.product.ProductDto;
 import com.flab.daitso.service.CategoryService;
-import com.flab.daitso.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/categories")
@@ -16,12 +13,9 @@ public class CategoryController {
 
     private final CategoryService categoryService;
 
-    private final ProductService productService;
-
     @Autowired
-    public CategoryController(CategoryService categoryService, ProductService productService) {
+    public CategoryController(CategoryService categoryService) {
         this.categoryService = categoryService;
-        this.productService = productService;
     }
 
     @PostMapping
@@ -34,34 +28,4 @@ public class CategoryController {
 //        return
 //        categoryService.findProductListByCategoryId(categoryId);
 //    }
-
-    @GetMapping("/{categoryId}")
-    public List<ProductDto> getProductListBySort(@PathVariable Long categoryId,
-                                                 @RequestParam(value = "score", required = false, defaultValue = "0") Long score,
-                                                 @RequestParam(value = "minPrice", required = false, defaultValue = "0") Long minPrice,
-                                                 @RequestParam(value = "maxPrice", required = false, defaultValue = "0") Long maxPrice,
-                                                 @RequestParam(value = "sorter", required = false, defaultValue = "0") Long sorter) {
-        if (score > 0) {
-            return getProductListByScoreRange(categoryId, score);
-        }
-        if (minPrice > 0 && (minPrice <= maxPrice)) {
-            return getProductListByPriceRange(categoryId, minPrice, maxPrice);
-        }
-        if (sorter >= 0) {
-            return getProductListByLatestOrder(categoryId, sorter);
-        }
-        return productService.findProductListByScoreRange(categoryId, 0L);
-    }
-
-    public List<ProductDto> getProductListByScoreRange(Long categoryId, Long score) {
-        return productService.findProductListByScoreRange(categoryId, score);
-    }
-
-    public List<ProductDto> getProductListByPriceRange(Long categoryId, Long minPrice, Long maxPrice) {
-        return productService.findProductListByPriceRange(categoryId, minPrice, maxPrice);
-    }
-
-    public List<ProductDto> getProductListByLatestOrder(Long categoryId, Long sorter) {
-        return productService.findProductListByLatestOrder(categoryId, sorter);
-    }
 }
